refactor(linker): isolate path safety operations
